+++ Tasks/task1/src/main/java/model/flowers/Tulip.java
@@ -9,7 +9,7 @@ import java.math.BigDecimal;
 public class Tulip extends Flower {
     private final static String name = "Tulip";
 
-    public Tulip(Color color, Recency recency, float stemLength, BigDecimal price) {
+    public Tulip(Color color, Recency recency, double stemLength, BigDecimal price) {
         super(name, color, recency, stemLength, price);
     }
 }
